@@ -62,7 +62,7 @@
     "@types/luxon": "^2.3.1",
     "@types/node": "^17.0.29",
     "@types/react": "^18.0.8",
-    "@types/react-dom": "18.0.2",
+    "@types/react-dom": "18.0.0",
     "@types/react-redux": "^7.1.24",
     "@types/redux-logger": "^3.0.9",
     "@typescript-eslint/eslint-plugin": "^5.21.0",
@@ -111,7 +111,7 @@
     "vite": "^2.9.6"
   },
   "peerDependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   }
 }
